Introducing initial support for compile-time / runtime output for PDF books

+++ frontend/tests/compile_blocks.py
@@ -30,6 +30,7 @@ import subprocess as S
 from os import path as P
 import colors as C
 import shutil
+import glob
 import re
 from widget.chop import manual_chop, cheapo_gnatchop, real_gnatchop
 
@@ -100,7 +101,10 @@ class Block(object):
                         i + 1,
                         lang_re.match(line).groups()[0]
                     )
-                    project = project_re.match(line).groups()[0]
+                    project = project_re.match(line)
+                    if project is not None:
+                        project = project.groups()[0]
+
                     main_file = main_re.match(line)
                     if main_file is not None:
                         # Retrieve actual main filename
@@ -211,6 +215,8 @@ parser.add_argument('--build-dir', '-B', type=str, default="build",
 parser.add_argument('--verbose', '-v', type=bool, default=False,
                     help='Show more information')
 
+parser.add_argument('--keep_files', '-k', action='store_true',
+                    help='Keep files generated in the test')
 parser.add_argument('--code-block', '-b', type=str, default=0)
 parser.add_argument('--all-diagnostics', '-A', action='store_true')
 parser.add_argument('--code-block-at', type=int, default=0)
@@ -344,6 +350,11 @@ def analyze_file(rst_file):
     projects = dict()
 
     for (i, b) in code_blocks:
+        if b.project is None:
+            print ("Error: project not set in {} at line {}".format(
+                rst_file, str(b.line_start)))
+            exit(1)
+
         if not b.project in projects:
             projects[b.project] = list()
         projects[b.project].append((i, b))
@@ -357,7 +368,7 @@ def analyze_file(rst_file):
             project_dir = base_project_dir + "/" + project.replace(".", "/")
 
             if os.path.exists(project_dir):
-                shutil.rmtree(args.build_dir)
+                shutil.rmtree(project_dir)
 
             try:
                 os.makedirs(project_dir)
@@ -454,6 +465,13 @@ def analyze_file(rst_file):
                     main_file = source_files[-1].basename
                 return main_file
 
+            def make_project_block_dir():
+                project_block_dir = str(block.line_start)
+                if not os.path.exists(project_block_dir):
+                    os.makedirs(project_block_dir)
+
+                return project_block_dir
+
             if (('ada-run' in block.classes
                  or 'ada-run-expect-failure' in block.classes
                  or 'run' in block.buttons)
@@ -461,7 +479,13 @@ def analyze_file(rst_file):
             ):
                 main_file = get_main_filename(block)
 
+                project_block_dir = make_project_block_dir()
+
                 if block.language == "ada":
+
+                    def remove_string(some_text, rem):
+                        return re.sub(".*" + rem + ".*\n?","", some_text)
+
                     try:
                         out = run("gprbuild", "-gnata", "-gnatyg0-s", "-f",
                                   main_file)
@@ -472,21 +496,32 @@ def analyze_file(rst_file):
                             print_error(loc, "Failed to compile example")
                             print(e.output)
                             has_error = True
+                        out = str(e.output.decode("utf-8"))
+
+                    out = remove_string(out, "using project")
+                    with open(project_block_dir + "/build.log", u"w") as logfile:
+                        logfile.write(out)
+
                 elif block.language == "c":
                     try:
-                        out = run("gcc", "-c", main_file)
+                        cmd = ["gcc", "-o",
+                               P.splitext(main_file)[0]] + glob.glob('*.c')
+                        out = run(*cmd)
                     except S.CalledProcessError as e:
-                        if 'ada-expect-compile-error' in block.classes:
+                        if 'c-expect-compile-error' in block.classes:
                             compile_error = True
                         else:
                             print_error(loc, "Failed to compile example")
                             print(e.output)
                             has_error = True
+                        out = str(e.output.decode("utf-8"))
+                    with open(project_block_dir + "/build.log", u"w") as logfile:
+                        logfile.write(out)
 
                 if not compile_error and not has_error:
                     if block.language == "ada":
                         try:
-                            run("./{}".format(P.splitext(main_file)[0]))
+                            out = run("./{}".format(P.splitext(main_file)[0]))
 
                             if 'ada-run-expect-failure' in block.classes:
                                 print_error(
@@ -494,7 +529,7 @@ def analyze_file(rst_file):
                                 )
                                 has_error = True
 
-                        except S.CalledProcessError:
+                        except S.CalledProcessError as e:
                             if 'ada-run-expect-failure' in block.classes:
                                 if args.verbose:
                                     print("Running of example expectedly failed")
@@ -502,23 +537,72 @@ def analyze_file(rst_file):
                                 print_error(loc, "Running of example failed")
                                 has_error = True
 
+                            out = str(e.output.decode("utf-8"))
+
+                        with open(project_block_dir + "/run.log", u"w") as logfile:
+                            logfile.write(out)
+
+                    elif block.language == "c":
+                        try:
+                            out = run("./{}".format(P.splitext(main_file)[0]))
+
+                            if 'c-run-expect-failure' in block.classes:
+                                print_error(
+                                    loc, "Running of example should have failed"
+                                )
+                                has_error = True
+
+                        except S.CalledProcessError as e:
+                            if 'c-run-expect-failure' in block.classes:
+                                if args.verbose:
+                                    print("Running of example expectedly failed")
+                            else:
+                                print_error(loc, "Running of example failed")
+                                has_error = True
+                            out = str(e.output.decode("utf-8"))
+
+                        with open(project_block_dir + "/run.log", u"w") as logfile:
+                            logfile.write(out)
+
             if 'compile' in block.buttons:
+
+                project_block_dir = make_project_block_dir()
 
                 for source_file in source_files:
                     if block.language == "ada":
                         try:
-                            run("gcc", "-c", "-gnatc", "-gnatyg0-s",
-                                source_file.basename)
-                        except S.CalledProcessError:
+                            out = run("gcc", "-c", "-gnatc", "-gnatyg0-s",
+                                      source_file.basename)
+                        except S.CalledProcessError as e:
                             if 'ada-expect-compile-error' in block.classes:
                                 compile_error = True
                             else:
                                 print_error(loc, "Failed to compile example")
                                 has_error = True
+                            out = str(e.output.decode("utf-8"))
+
+                        with open(project_block_dir + "/compile.log", u"w+") as logfile:
+                            logfile.write(out)
+
+                    elif block.language == "c":
+                        try:
+                            out = run("gcc", "-c", source_file.basename)
+                        except S.CalledProcessError as e:
+                            if 'c-expect-compile-error' in block.classes:
+                                compile_error = True
+                            else:
+                                print_error(loc, "Failed to compile example")
+                                has_error = True
+                            out = str(e.output.decode("utf-8"))
+
+                        with open(project_block_dir + "/compile.log", u"w+") as logfile:
+                            logfile.write(out)
 
             if any(b in prove_buttons for b in block.buttons):
 
                 if block.language == "ada":
+                    project_block_dir = make_project_block_dir()
+
                     main_file = get_main_filename(block)
                     spark_mode = True
                     project_filename = write_project_file(main_file,
@@ -552,6 +636,10 @@ def analyze_file(rst_file):
                             print_error(loc, "Failed to prove example")
                             print(e.output)
                             has_error = True
+                        out = str(e.output.decode("utf-8"))
+
+                    with open(project_block_dir + "/prove.log", u"w") as logfile:
+                        logfile.write(out)
                 else:
                     print_error(loc, "Wrong language selected for prove button")
                     print(e.output)
@@ -579,7 +667,7 @@ def analyze_file(rst_file):
                     print_error(loc, "Expected prove error, got none!")
                     has_error = True
 
-            if (any (c in ['ada-run','ada-run-expect-failure','ada-norun'] for
+            if (any (c in ['ada-run-expect-failure','ada-norun'] for
                      c in block.classes)
                 and not 'run' in block.buttons):
                 print_error(loc, "Expected run button, got none!")
@@ -595,14 +683,15 @@ def analyze_file(rst_file):
 
         os.chdir(work_dir)
 
-    if os.path.exists(base_project_dir):
+    if os.path.exists(base_project_dir) and not args.keep_files:
         shutil.rmtree(base_project_dir)
 
     return analysis_error
 
 # Remove the build dir, but only if the user didn't ask for a specific
 # subset of code_blocks
-if os.path.exists(args.build_dir) and not args.code_block:
+if (os.path.exists(args.build_dir) and not args.code_block
+    and not args.keep_files):
     shutil.rmtree(args.build_dir)
 
 if not os.path.exists(args.build_dir):
